backend/researcher/pyproject.toml
[project]
name = "researcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "openai-agents[litellm]>=0.2.4",
    "playwright>=1.54.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "tenacity>=9.1.2",
    "uvicorn>=0.35.0",
]
